Fixed couple ITs
git-svn-id: file:///opt/svn/repositories/sonatype.org/nexus/trunk@7048 2aa8b3fc-8ebb-4439-a84f-95066eaea8ab

+++ nexus/nexus-test-harness/nexus-test-harness-launcher/src/main/java/org/sonatype/nexus/test/utils/NexusStatusUtil.java
@@ -127,8 +127,14 @@ public class NexusStatusUtil
         throw new NexusIllegalStateException( "Unable to doHardStart(), nexus still stopped, took 200s" );
 
     }
-
+    
     public void stop()
+        throws Exception
+    {
+        stop( true );
+    }
+
+    public void stop( boolean deleteTasks )
         throws Exception
     {
         if ( APP_BOOTER_SERVICE == null )
@@ -143,7 +149,14 @@ public class NexusStatusUtil
         {
             // NOTE: We can't kill active tasks, we need to wait for them to stop
             TestContainer.getInstance().getTestContext().useAdminForRequests();
-            TaskScheduleUtil.deleteAllTasks();
+            if ( deleteTasks )
+            {
+                TaskScheduleUtil.deleteAllTasks();
+            }
+            else
+            {
+                TaskScheduleUtil.waitForAllTasksToStop();
+            }
 
             try
             {
